gateway/src/main/java/com/hfarhani/gateway/service/JwtUserDetailsService.java
package com.hfarhani.gateway.service;

import com.hfarhani.gateway.dto.UserDto;
import com.hfarhani.gateway.model.JpaUser;
import com.hfarhani.gateway.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class JwtUserDetailsService implements UserDetailsService {

    @Autowired
    UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        JpaUser user = userRepository.findUserByUsername(username);
        if (user == null) {
            throw new UsernameNotFoundException("User not found with username: " + username);
        }
        UserDto userDto = user.convertToDto();
        return userDto;
    }

}
